Replaced picker-network components (typescript)
