--- ui/components/component-library/picker-network/README.mdx
@@ -1,48 +1,0 @@
-import { Story, Canvas, ArgsTable } from '@storybook/addon-docs';
-
-import { PickerNetwork } from './picker-network';
-
-# PickerNetwork
-
-The `PickerNetwork` is used for the action of changing a network.
-
-<Canvas>
-  <Story id="components-componentlibrary-pickernetwork--default-story" />
-</Canvas>
-
-## Props
-
-The `PickerNetwork` accepts all props below as well as all [Box](/docs/components-ui-box--default-story#props) component props.
-
-<ArgsTable of={PickerNetwork} />
-
-### Label
-
-Use the `label` prop for the text content of the `PickerNetwork` component. For long labels set a `max-width` using a `className` and the text will truncate showing an ellipsis.
-
-<Canvas>
-  <Story id="components-componentlibrary-pickernetwork--label" />
-</Canvas>
-
-```jsx
-import { PickerNetwork } from '../../ui/component-library';
-<PickerNetwork label="Arbitrum One" />
-<PickerNetwork label="Polygon Mainnet" />
-<PickerNetwork label="Optimism" />
-<PickerNetwork label="BNB Smart Chain (previously Binance Smart Chain Mainnet)" style={{ maxWidth: '200px' }} />
-```
-
-### Src
-
-Use the `src` prop with an image url to render the `AvatarNetwork`. Use the `avatarNetworkProps` to pass additional props to the `AvatarNetwork` component.
-
-<Canvas>
-  <Story id="components-componentlibrary-pickernetwork--src" />
-</Canvas>
-
-```jsx
-import { PickerNetwork } from '../../ui/component-library';
-<PickerNetwork src="./images/arbitrum.svg" />
-<PickerNetwork src="./images/matic-token.png" />
-<PickerNetwork src="./images/optimism.svg" />
-```
