--- ui/components/component-library/picker-network/index.js
@@ -1,1 +1,0 @@
-export { PickerNetwork } from './picker-network';
